agent-system/requirements.txt
langchain
langchain-groq
langgraph
fastmcp
python-dotenv
pydantic